+++ src/fontra/client/core/parse-glyphset.js
